Move `StyleManagerDelegate` and `Date` extension to a separate files.
